Cleaned up code with tools pycodestyle and pydocstyle to conform to PEP8 and PEP257, also, implemented getters and setters with properties.
This means basically that the code has been cleaned to conform to the
Python conventions. These are based on readability and consistensy. For
more information read 'The Zen of Python'

NEXT: Work on cell actions to be able co seamlessly couple a logger to
track the system's state and be able to reconstruct an inheritance tree.